Added support for tap to click and two finger scrolling
Galaxy Tab cover keyboard seems to use special Motion Event flags and not report more than one finger when doing two finger scrolling. We intercept this behaviour and transmit scroll events.

Tap to click is also a feature you can activate on the Tab keyboard, however it doesn't report itself as "PRIMARY" mouse button, which is confusing, probably to retain compatibility with touch focused apps. Added a path to handle this.

This currently breaks touch input however, and two finger scrolling currently is not working when not treating touch screen as touch pad.

Also fixed mouse cursor "hover" support, where the mouse cursor would not move in response to touchpad swipes. We now treat it the same as the S-Pen